requirements.txt
# Streamlit
streamlit==1.39.0
streamlit-folium==0.15.1     # Mapy interaktywne

# Geospatial - Wektorowe
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
fiona==1.9.5

# Geospatial - Rastrowe
rasterio==1.3.9
rioxarray==0.15.0
xarray==2023.12.0
rasterstats==0.19.0          # ⭐ DODANE - zonal statistics

# GDAL (zainstalowane przez base image)
# gdal==3.8.0  # Nie instaluj przez pip!

# Copernicus/Satellite
sentinelsat==1.2.1           # Pobieranie danych Sentinel
pystac-client==0.7.5         # STAC API

# Obliczenia i wizualizacja
numpy==1.26.2
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0
Pillow==10.1.0

# Inne
requests==2.31.0
